requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
gunicorn==21.2.0
pandas==2.1.4
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.4
google-cloud-vision==3.7.4
google-generativeai==0.8.4
gspread==6.1.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-auth==2.27.0
